Move to percolator, and reduce unnecessary public visibility
