+++ modules/percolator/src/main/java/org/elasticsearch/percolator/PercolatorFieldMapper.java
@@ -7,7 +7,6 @@
  */
 package org.elasticsearch.percolator;
 
-import org.elasticsearch.lucene.document.BinaryRange;
 import org.apache.lucene.document.Field;
 import org.apache.lucene.document.FieldType;
 import org.apache.lucene.document.NumericDocValuesField;
