+++ modules/percolator/src/main/java/org/elasticsearch/percolator/QueryAnalyzer.java
@@ -7,7 +7,6 @@
  */
 package org.elasticsearch.percolator;
 
-import org.elasticsearch.lucene.document.BinaryRange;
 import org.apache.lucene.index.PrefixCodedTerms;
 import org.apache.lucene.index.Term;
 import org.apache.lucene.queries.BlendedTermQuery;
